Replaced the featureMask type with uint8_t

+++ engine/src/Importer.cpp
@@ -458,7 +458,7 @@ struct ShadeHeader
     char     assetID            = (char)EFileType::SHADER;
     int      vertexPathLenght   = 0;
     int      fragmentPathLenght = 0;
-    uint16_t featureMask        = 0;
+    uint8_t  featureMask        = 0;
 };
 
 void GPE::writeShaderFile(const char* dst, const ShaderCreateonfig& arg)
